douyin_live_chat_recv/requirements.txt
brotli
nodejs-bin
protobuf
requests
websocket-client